API-892: Rework service to command/handler + rename sql query
