+++ src/Akeneo/Apps/back/Domain/Audit/Persistence/Query/ExtractAppsProductEventCountQuery.php
@@ -9,7 +9,7 @@ namespace Akeneo\Apps\Domain\Audit\Persistence\Query;
  * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
  * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-interface ExtractAppsEventCountQuery
+interface ExtractAppsProductEventCountQuery
 {
     public function extractCreatedProducts(string $date): array;
