+++ src/Akeneo/Apps/back/Infrastructure/Symfony/DependencyInjection/AkeneoAppsExtension.php
@@ -23,7 +23,6 @@ class AkeneoAppsExtension extends Extension
         $loader->load('validators.yml');
 
         $loader->load('cli.yml');
-        $loader->load('app_services.yml');
         $loader->load('handlers.yml');
         $loader->load('queries.yml');
         $loader->load('repositories.yml');
