--- src/Akeneo/Apps/back/Infrastructure/Symfony/Resources/config/app_services.yml
@@ -1,6 +1,0 @@
-services:
-    akeneo_app.application.audit.service.update_product_event_count:
-        class: Akeneo\Apps\Application\Audit\Service\UpdateProductEventCountService
-        arguments:
-            - '@akeneo_app.persistence.query.extract_apps_event_count'
-            - '@akeneo_app.persistence.repository.event_count'
