+++ src/Akeneo/Apps/back/Infrastructure/Symfony/Resources/config/queries.yml
@@ -14,8 +14,8 @@ services:
         arguments:
             - '@database_connection'
 
-    akeneo_app.persistence.query.extract_apps_event_count:
-        class: Akeneo\Apps\Infrastructure\Persistence\Dbal\Query\DbalExtractAppsEventCountQuery
+    akeneo_app.persistence.query.extract_apps_product_event_count:
+        class: Akeneo\Apps\Infrastructure\Persistence\Dbal\Query\DbalExtractAppsProductEventCountQuery
         arguments:
             - '@database_connection'
             - '%pim_catalog.entity.product.class%'
