+++ src/Akeneo/Apps/back/tests/Unit/spec/Application/Audit/Command/UpdateProductEventCountHandlerSpec.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace spec\Akeneo\Apps\Application\Audit\Service;
+namespace spec\Akeneo\Apps\Application\Audit\Command;
 
-use Akeneo\Apps\Application\Audit\Service\UpdateProductEventCountService;
-use Akeneo\Apps\Domain\Audit\Persistence\Query\ExtractAppsEventCountQuery;
+use Akeneo\Apps\Application\Audit\Command\UpdateProductEventCountHandler;
+use Akeneo\Apps\Domain\Audit\Persistence\Query\ExtractAppsProductEventCountQuery;
 use Akeneo\Apps\Domain\Audit\Persistence\Repository\EventCountRepository;
 use PhpSpec\ObjectBehavior;
 
@@ -14,15 +14,15 @@ use PhpSpec\ObjectBehavior;
  * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
  * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class UpdateProductEventCountServiceSpec extends ObjectBehavior
+class UpdateProductEventCountHandlerSpec extends ObjectBehavior
 {
-    function let(ExtractAppsEventCountQuery $extractAppsEventCountQuery, EventCountRepository $eventCountRepository)
+    function let(ExtractAppsProductEventCountQuery $extractAppsEventCountQuery, EventCountRepository $eventCountRepository)
     {
         $this->beConstructedWith($extractAppsEventCountQuery, $eventCountRepository);
     }
 
     function it_is_initializable()
     {
-        $this->shouldBeAnInstanceOf(UpdateProductEventCountService::class);
+        $this->shouldBeAnInstanceOf(UpdateProductEventCountHandler::class);
     }
 }
